fix inherits bug and default date values
git-svn-id: http://svn.php.net/repository/pear/packages/Gtk_MDB_Designer/trunk@120973 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Designer.php
@@ -174,8 +174,9 @@ class Gtk_MDB_Designer {
     */   
     function loadFile($file) {
     
-
-        $this->database->destroy();
+        if ($this->database) {
+            $this->database->destroy();
+        }
         // load a new parser.
         
         $parser = &new Gtk_MDB_Designer_Parser;
